Move deduplication code to DuplicationFilter
